python-backend/requirements.txt
Flask>=2.3.0
pdfplumber>=0.11.0
scikit-learn>=1.4.0
numpy>=1.26.0
groq>=0.9.0
openai>=1.30.0
python-dotenv>=1.0.0
